agentic/requirements.txt
google-adk
uvicorn
fastapi
dotenv
pydantic[email]
